+++ test/tests.py
@@ -1,16 +1,19 @@
+import gzip
+import logging
 import os
 import os.path
 import tempfile
 import unittest
-import logging
+from unittest.mock import Mock
+
+from flask import Flask, Response, json
+from flask_restplus import Resource, Api
 from pycommon_test.flask_restplus_mock import TestAPI
 from pycommon_test.samba_mock import TestConnection
 from pycommon_test.service_tester import JSONTestCase
-from flask import Flask
-from flask_restplus import Resource, Api
 
-from pycommon_server.configuration import load_configuration, load_logging_configuration, load
 from pycommon_server import flask_restplus_common, logging_filter, windows
+from pycommon_server.configuration import load_configuration, load_logging_configuration, load
 
 logger = logging.getLogger(__name__)
 
@@ -157,6 +160,160 @@ class ConfigurationTest(unittest.TestCase):
                 load(os.path.join(server_folder, 'server.py')))
 
 
+class HealthCheckWithException(JSONTestCase):
+    def create_app(self):
+        app = Flask(__name__)
+        app.config['TESTING'] = True
+        api = Api(app, version='3.2.1')
+
+        def throw_exception():
+            raise Exception('This is the error message.')
+
+        flask_restplus_common.add_monitoring_namespace(api, throw_exception)
+
+        return app
+
+    def test_health_check_response_on_exception(self):
+        response = self.client.get('/health')
+        self.assert400(response)
+        self.assert_json(response, {
+            'details': {},
+            'output': 'This is the error message.',
+            'releaseId': '3.2.1',
+            'status': 'fail',
+            'version': '3',
+        })
+
+
+class HealthCheckWithFailureDetails(JSONTestCase):
+    def create_app(self):
+        app = Flask(__name__)
+        app.config['TESTING'] = True
+        api = Api(app, version='3.2.1')
+
+        def failure_details():
+            return None, {'toto': {'status': 'warn'}}, {'toto2': {'status': 'fail'}}
+
+        flask_restplus_common.add_monitoring_namespace(api, failure_details)
+
+        return app
+
+    def test_health_check_response_on_exception(self):
+        response = self.client.get('/health')
+        self.assert400(response)
+        self.assert_json(response, {
+            'details': {'toto': {'status': 'warn'}, 'toto2': {'status': 'fail'}},
+            'releaseId': '3.2.1',
+            'status': 'fail',
+            'version': '3',
+        })
+
+
+class HealthCheckWithWarnDetails(JSONTestCase):
+    def create_app(self):
+        app = Flask(__name__)
+        app.config['TESTING'] = True
+        api = Api(app, version='3.2.1')
+
+        def warning_details():
+            return {'toto2': {'status': 'pass'}}, {'toto': {'status': 'warn'}}, None
+
+        flask_restplus_common.add_monitoring_namespace(api, warning_details)
+
+        return app
+
+    def test_health_check_response_on_warning(self):
+        response = self.client.get('/health')
+        self.assert200(response)
+        self.assert_json(response, {
+            'details': {'toto2': {'status': 'pass'}, 'toto': {'status': 'warn'}},
+            'releaseId': '3.2.1',
+            'status': 'warn',
+            'version': '3',
+        })
+
+
+class HealthCheckWithPassDetails(JSONTestCase):
+    def create_app(self):
+        app = Flask(__name__)
+        app.config['TESTING'] = True
+        api = Api(app, version='3.2.1')
+
+        def pass_details():
+            return {'toto2': {'status': 'pass'}}, None, None
+
+        flask_restplus_common.add_monitoring_namespace(api, pass_details)
+
+        return app
+
+    def test_health_check_response_on_pass(self):
+        response = self.client.get('/health')
+        self.assert200(response)
+        self.assert_json(response, {
+            'details': {'toto2': {'status': 'pass'}},
+            'releaseId': '3.2.1',
+            'status': 'pass',
+            'version': '3',
+        })
+
+    def test_generated_swagger(self):
+        response = self.client.get('/swagger.json')
+        self.assert200(response)
+        self.assert_swagger(response, {
+            'swagger': '2.0',
+            'basePath': '/',
+            'paths': {
+                '/health': {
+                    'get': {
+                        'responses': {
+                            '200': {
+                                'description': 'Server is in a coherent state.',
+                                'schema': {'$ref': '#/definitions/HealthPass'}
+                            },
+                            '400': {
+                                'description': 'Server is not in a coherent state.',
+                                'schema': {'$ref': '#/definitions/HealthFail'}
+                            }
+                        },
+                        'summary': 'Check service health',
+                        'description': 'This endpoint perform a quick server state check.',
+                        'operationId': 'get_health',
+                        'tags': ['monitoring']
+                    }
+                }
+            },
+            'info': {'title': 'API', 'version': '3.2.1'},
+            'produces': ['application/json'],
+            'consumes': ['application/json'],
+            'tags': [{'name': 'monitoring', 'description': 'Monitoring operations'}],
+            'definitions': {
+                'HealthPass': {
+                    'required': ['details', 'releaseId', 'status', 'version'],
+                    'properties': {
+                        'status': {'type': 'string', 'description': 'Indicates whether the service status is acceptable or not.', 'example': 'pass', 'enum': ['pass', 'warn']},
+                        'version': {'type': 'string', 'description': 'Public version of the service.', 'example': '1'},
+                        'releaseId': {'type': 'string', 'description': 'Version of the service.', 'example': '1.0.0'},
+                        'details': {'type': 'object', 'description': 'Provides more details about the status of the service.'}
+                    }, 'type': 'object'
+                },
+                'HealthFail': {
+                    'required': ['details', 'releaseId', 'status', 'version'],
+                    'properties': {
+                        'status': {'type': 'string', 'description': 'Indicates whether the service status is acceptable or not.', 'example': 'fail', 'enum': ['fail']},
+                        'version': {'type': 'string', 'description': 'Public version of the service.', 'example': '1'},
+                        'releaseId': {'type': 'string', 'description': 'Version of the service.', 'example': '1.0.0'},
+                        'details': {'type': 'object', 'description': 'Provides more details about the status of the service.'},
+                        'output': {'type': 'string', 'description': 'Raw error output.'}
+                    }, 'type': 'object'
+                }
+            },
+            'responses': {
+                'ParseError': {'description': "When a mask can't be parsed"},
+                'MaskError': {'description': 'When any error occurs on mask'}
+            }
+        })
+
+
 class FlaskRestPlusTest(JSONTestCase):
     def create_app(self):
         app = Flask(__name__)
@@ -199,21 +356,92 @@ class FlaskRestPlusTest(JSONTestCase):
             def delete(self):
                 return ''
 
+        @api.route('/standard_responses')
+        class StandardResponses(Resource):
+            @api.doc(**flask_restplus_common.created_response_doc(api))
+            def post(self):
+                return flask_restplus_common.created_response('/standard_responses?id=42')
+
+            @api.doc(**flask_restplus_common.updated_response_doc(api))
+            def put(self):
+                return flask_restplus_common.updated_response('/standard_responses?id=43')
+
+            @api.response(*flask_restplus_common.deleted_response_doc)
+            def delete(self):
+                return flask_restplus_common.deleted_response
+
         return app
 
-    def test_successful_return(self):
-        self.assertEqual(({'status': 'Successful'}, 200), flask_restplus_common.successful_return)
+    def test_standard_post_response_without_reverse_proxy(self):
+        response = self.post_json('/standard_responses', {})
+        self.assertStatus(response, 201)
+        self.assert_json(response, {'status': 'Successful'})
+        self.assertEqual(response.headers['location'], 'http://localhost/standard_responses?id=42')
 
-    def test_successful_deletion_return(self):
-        self.assertEqual(('', 204), flask_restplus_common.successful_deletion_return)
+    def test_standard_post_response_with_reverse_proxy(self):
+        response = self.post_json('/standard_responses', {}, headers={
+            'X-Original-Request-Uri': '/reverse/standard_responses',
+            'Host': 'localhost',
+        })
+        self.assertStatus(response, 201)
+        self.assert_json(response, {'status': 'Successful'})
+        self.assertEqual(response.headers['location'], 'http://localhost/reverse/standard_responses?id=42')
 
-    def test_successful_deletion_response(self):
-        self.assertEqual((204, 'Sample deleted'), flask_restplus_common.successful_deletion_response)
+    def test_standard_put_response_without_reverse_proxy(self):
+        response = self.put_json('/standard_responses', {})
+        self.assertStatus(response, 201)
+        self.assert_json(response, {'status': 'Successful'})
+        self.assertEqual(response.headers['location'], 'http://localhost/standard_responses?id=43')
 
-    def test_successful_model(self):
-        model = flask_restplus_common.successful_model(TestAPI)
-        self.assertEqual('Successful', model.name)
-        self.assertEqual({'status': 'Successful'}, model.fields_default)
+    def test_standard_put_response_with_reverse_proxy(self):
+        response = self.put_json('/standard_responses', {}, headers={
+            'X-Original-Request-Uri': '/reverse/standard_responses',
+            'Host': 'localhost',
+        })
+        self.assertStatus(response, 201)
+        self.assert_json(response, {'status': 'Successful'})
+        self.assertEqual(response.headers['location'], 'http://localhost/reverse/standard_responses?id=43')
+
+    def test_standard_delete_response(self):
+        response = self.client.delete('/standard_responses')
+        self.assertStatus(response, 204)
+
+    def test_generated_swagger(self):
+        response = self.client.get('/swagger.json')
+        self.assert200(response)
+        self.assert_swagger(response, {
+            'swagger': '2.0', 'basePath': '/', 'paths': {
+                '/logging': {
+                    'delete': {'responses': {'200': {'description': 'Success'}}, 'operationId': 'delete_logging', 'tags': ['default']},
+                    'get': {'responses': {'200': {'description': 'Success'}}, 'operationId': 'get_logging', 'tags': ['default']},
+                    'post': {'responses': {'200': {'description': 'Success'}}, 'operationId': 'post_logging', 'tags': ['default']},
+                    'put': {'responses': {'200': {'description': 'Success'}}, 'operationId': 'put_logging', 'tags': ['default']}
+                },
+                '/requires_authentication': {
+                    'delete': {'responses': {'200': {'description': 'Success'}}, 'operationId': 'delete_requires_authentication', 'tags': ['default']},
+                    'get': {'responses': {'200': {'description': 'Success'}}, 'operationId': 'get_requires_authentication', 'tags': ['default']},
+                    'post': {'responses': {'200': {'description': 'Success'}}, 'operationId': 'post_requires_authentication', 'tags': ['default']},
+                    'put': {'responses': {'200': {'description': 'Success'}}, 'operationId': 'put_requires_authentication', 'tags': ['default']}
+                },
+                '/standard_responses': {
+                    'delete': {'responses': {'204': {'description': 'Deleted'}}, 'operationId': 'delete_standard_responses', 'tags': ['default']},
+                    'post': {'responses': {'201': {'description': 'Created', 'headers': {'location': {'description': 'Location of created resource.', 'type': 'string'}}, 'schema': {'$ref': '#/definitions/Created'}}}, 'operationId': 'post_standard_responses', 'tags': ['default']},
+                    'put': {'responses': {'201': {'description': 'Updated', 'headers': {'location': {'description': 'Location of updated resource.', 'type': 'string'}}, 'schema': {'$ref': '#/definitions/Updated'}}}, 'operationId': 'put_standard_responses', 'tags': ['default']}
+                }
+            },
+            'info': {'title': 'API', 'version': '1.0'},
+            'produces': ['application/json'],
+            'consumes': ['application/json'],
+            'tags': [{'name': 'default', 'description': 'Default namespace'}],
+            'responses': {
+                'ParseError': {'description': "When a mask can't be parsed"},
+                'MaskError': {'description': 'When any error occurs on mask'}
+            },
+            'definitions': {
+                'Created': {'properties': {'status': {'default': 'Successful', 'type': 'string'}}, 'type': 'object'},
+                'Updated': {'properties': {'status': {'default': 'Successful', 'type': 'string'}}, 'type': 'object'},
+            },
+        })
 
     def test_authentication_failure_token_not_provided_on_get(self):
         response = self.client.get('/requires_authentication')
@@ -513,7 +741,7 @@ class WindowsTest(unittest.TestCase):
         with self.assertRaises(FileNotFoundError) as cm:
             windows.rename(connection, 'TestShare\\', 'file_to_rename_2', 'file_new_name')
 
-        self.assertEqual(str(cm.exception),  r"\\TestComputer\TestShare\file_to_rename_2 doesn't exist")
+        self.assertEqual(str(cm.exception), r"\\TestComputer\TestShare\file_to_rename_2 doesn't exist")
 
     def test_get_file_desc_file_exists(self):
         connection = windows.connect('TestComputer', '127.0.0.1', 80, 'TestDomain', 'TestUser', 'TestPassword')
@@ -532,6 +760,94 @@ class WindowsTest(unittest.TestCase):
         founded_file = windows.get_file_desc(connection, 'TestShare/', 'nonexistent_file_to_find')
 
         self.assertIsNone(founded_file)
+
+
+class CreateNewApi(unittest.TestCase):
+    def test_basic_api(self):
+        import test.test_get_user as fake_service
+        app, api = flask_restplus_common.create_api(fake_service.__name__, title='TestApi', description='Testing API', cors=False,
+                                                    reverse_proxy=False)
+
+        with app.test_client() as client:
+            response = client.get('/swagger.json')
+            JSONTestCase().assert_200(response)
+            JSONTestCase().assert_json(response, {'swagger': '2.0', 'basePath': '/', 'paths': {},
+                                                  'info': {'title': 'TestApi', 'version': '1.0.0',
+                                                           'description': 'Testing API'},
+                                                  'produces': ['application/json'], 'consumes': ['application/json'],
+                                                  'tags': [], 'responses': {
+                    'ParseError': {'description': "When a mask can't be parsed"},
+                    'MaskError': {'description': 'When any error occurs on mask'}}})
+
+    def test_cors_api(self):
+        import test.test_get_user as fake_service
+        app, api = flask_restplus_common.create_api(fake_service.__name__, title='TestApi', description='Testing API',
+                                                    reverse_proxy=False)
+
+        with app.test_client() as client:
+            response = client.get('/swagger.json')
+            JSONTestCase().assert_200(response)
+            JSONTestCase().assert_json(response, {'swagger': '2.0', 'basePath': '/', 'paths': {},
+                                                  'info': {'title': 'TestApi', 'version': '1.0.0',
+                                                           'description': 'Testing API'},
+                                                  'produces': ['application/json'], 'consumes': ['application/json'],
+                                                  'tags': [], 'responses': {
+                    'ParseError': {'description': "When a mask can't be parsed"},
+                    'MaskError': {'description': 'When any error occurs on mask'}}})
+            self.assertEqual(response.headers.get('Access-Control-Allow-Origin'), '*')
+
+    def test_compress_api(self):
+        import test.test_get_user as fake_service
+        app, api = flask_restplus_common.create_api(fake_service.__name__, title='TestApi', description='Testing API', cors=False,
+                                                    reverse_proxy=False,
+                                                    compress_mimetypes=['application/json'])
+
+        heavy_answer = {'test': 1000 * 'A'}
+
+        @api.route('/test')
+        class TestRoute(Resource):
+            def get(self):
+                return Response(response=json.dumps(heavy_answer), mimetype='application/json')
+
+        with app.test_client() as client:
+            response = client.get('/test', headers=[('Accept-Encoding', 'gzip')])
+            JSONTestCase().assert_200(response)
+            self.assertEqual(response.content_encoding, 'gzip')
+            mock_response = Mock(data=gzip.decompress(response.data))
+            JSONTestCase().assert_json(mock_response, heavy_answer)
+
+    def test_reverse_proxy_api(self):
+        import test.test_get_user as fake_service
+        app, api = flask_restplus_common.create_api(fake_service.__name__, title='TestApi', description='Testing API', cors=False)
+
+        with app.test_client() as client:
+            response = client.get('/swagger.json', headers=[('X-Original-Request-Uri', '/behind_reverse_proxy')])
+            JSONTestCase().assert_200(response)
+            JSONTestCase().assert_json(response, {'swagger': '2.0', 'basePath': '/behind_reverse_proxy', 'paths': {},
+                                                  'info': {'title': 'TestApi', 'version': '1.0.0',
+                                                           'description': 'Testing API'},
+                                                  'produces': ['application/json'], 'consumes': ['application/json'],
+                                                  'tags': [], 'responses': {
+                    'ParseError': {'description': "When a mask can't be parsed"},
+                    'MaskError': {'description': 'When any error occurs on mask'}}})
+
+    def test_extra_parameters_api(self):
+        import test.test_get_user as fake_service
+        app, api = flask_restplus_common.create_api(fake_service.__name__, title='TestApi', description='Testing API', cors=False,
+                                                    reverse_proxy=False, license_url='engie.license.com',
+                                                    license='engie')
+
+        with app.test_client() as client:
+            response = client.get('/swagger.json', headers=[('X-Original-Request-Uri', '/behind_reverse_proxy')])
+            JSONTestCase().assert_200(response)
+            JSONTestCase().assert_json(response, {'swagger': '2.0', 'basePath': '/', 'paths': {},
+                                                  'info': {'title': 'TestApi', 'version': '1.0.0',
+                                                           'description': 'Testing API',
+                                                           'license': {'name': 'engie', 'url': 'engie.license.com'}},
+                                                  'produces': ['application/json'], 'consumes': ['application/json'],
+                                                  'tags': [], 'responses': {
+                    'ParseError': {'description': "When a mask can't be parsed"},
+                    'MaskError': {'description': 'When any error occurs on mask'}}})
 
 
 if __name__ == '__main__':